The Titanic Datasets
The training set should be used to build your machine learning models. For the training set, we provide the outcome (also known as the “ground truth”) for each passenger.

The test set should be used to see how well your model performs on unseen data. For the test set, we do not provide the ground truth for each passenger.